+++ src/rsdict/__init__.py
@@ -1,3 +1,3 @@
 from .rsdict import rsdict, rsdict_frozen, rsdict_unfix, rsdict_fixtype, rsdict_fixkey
 
-__version__ = "0.1.4"
+__version__ = "0.1.5"