+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 **Merged pull requests:**
 -->
 
+## v0.1.5
+
+- Add typechecking for initial arguments.
+- Add optional argument: `get_initval(key)`.
+
 ## v0.1.4
 
 - Add subclasses `rsdict_frozen`, `rsdict_unfix`, `rsdict_fixkey`, `rsdict_fixtype`.